backend/pyproject.toml
[project]
name = "role-based-rag-chatbot"
version = "0.1.0"
description = "Role-based RAG chatbot using FastAPI, vector search, and YAML-driven access control"
readme = "README.md"
requires-python = ">=3.10,<=3.13"

dependencies = [
  # API
  "fastapi",
  "uvicorn[standard]",
  # RAG / NLP
  "langchain",
  "langchain-community",
  "langchain-core",
  "langchain-text-splitters",
  "langchain-huggingface",
  "chromadb",
  "sentence-transformers",
  # Data processing
  "pandas>=2.1",
  "pypdf",
  "python-docx",
  "unstructured",
  "markdown",
  # Auth & Config
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "python-multipart",
  "pydantic-settings",
  "pyyaml",
  "langchain-groq>=1.1.1",
  "langsmith>=0.6.3",
]
